--- a/src/NuGetGallery.Core/Entities/Organization.cs
+++ b/src/NuGetGallery.Core/Entities/Organization.cs
@@ -1,7 +1,10 @@
 ﻿// Copyright (c) .NET Foundation. All rights reserved.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
+using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
+using System.Linq;
 
 namespace NuGetGallery
 {
@@ -15,14 +18,18 @@ namespace NuGetGallery
     /// <see href="https://weblogs.asp.net/manavi/inheritance-mapping-strategies-with-entity-framework-code-first-ctp5-part-2-table-per-type-tpt" />
     public class Organization : User
     {
-        public Organization() : base()
+        public Organization() : this(null)
         {
-            Members = new List<Membership>();
         }
 
         public Organization(string name) : base(name)
         {
             Members = new List<Membership>();
+
+            _administrators = new Lazy<IEnumerable<User>>(
+                () => Members.Where(m => m.IsAdmin).Select(m => m.Member).ToList());
+            _collaborators = new Lazy<IEnumerable<User>>(
+                () => Members.Where(m => !m.IsAdmin).Select(m => m.Member).ToList());
         }
 
         /// <summary>
@@ -34,5 +41,31 @@ namespace NuGetGallery
         /// Requests to become a member of this <see cref="Organization"/>.
         /// </summary>
         public virtual ICollection<MembershipRequest> MemberRequests { get; set; }
+
+        #region per-request query cache
+
+        private Lazy<IEnumerable<User>> _administrators;
+        private Lazy<IEnumerable<User>> _collaborators;
+
+        [NotMapped]
+        public IEnumerable<User> Administrators
+        {
+            get
+            {
+                return _administrators.Value;
+            }
+        }
+
+        [NotMapped]
+        public IEnumerable<User> Collaborators
+        {
+            get
+            {
+                return _collaborators.Value;
+            }
+        }
+
+        #endregion
+
     }
 }