--- a/src/NuGetGallery/Services/UserService.cs
+++ b/src/NuGetGallery/Services/UserService.cs
@@ -109,7 +109,7 @@ namespace NuGetGallery
             if (membership.IsAdmin != isAdmin)
             {
                 // block removal of last admin
-                if (membership.IsAdmin && organization.Members.Count(m => m.IsAdmin) == 1)
+                if (membership.IsAdmin && organization.Administrators.Count() == 1)
                 {
                     throw new EntityException(string.Format(CultureInfo.CurrentCulture,
                         Strings.UpdateOrDeleteMember_CannotRemoveLastAdmin, memberName));