Issue #5461, optimization 2.

Caches and separates admin/collaborators lookups for current user during permissions checks so duplicate checks don't hit the DB.

Originally saw response go from 37.8s -> 35.1s, but now seeing it go from 59.1s -> 42.5s